feat：新增业务

+++ demo/src/main/java/com/lin/cms/demo/sleeve/model/Activity.java
@@ -49,5 +49,5 @@ public class Activity implements Serializable {
 
     private String name;
 
-    private Integer activityCoverId;
+    private Long activityCoverId;
 }
